draft_preparation/multi_backend_project/servers/collab-python/pyproject.toml
[project]
name = "collab-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Harshith Kurapati",email = "harshith.kurapati@ovaledge.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "redis (>=6.4.0,<7.0.0)",
]

[tool.poetry]
packages = [{include = "collab_python", from = "src"}]


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.37.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "collab_python.main:start"